+++ releases/ChangeLog.md
@@ -3,6 +3,8 @@
 - Bugfix: Saving passphrase on SD Card caused a freeze that required reboot
 - Bugfix: Properly handle and finalize framing error response
 - Bugfix: `Brick Me` option for `If Wrong` PIN caused yikes
+- Bugfix: Do not allow non-ascii or ascii non-printable characters in multisig
+  wallet name
 
 ## 5.2.2 - 2023-12-21